Agregada pagina Panel (panel de control)
- En esta pagina se pueden modificar los datos del usuario, con la
excepcion del email, password (por ahora), el id, tipo de usuario y estado.
- Se muestran los datos, junto con la imagen, con la opcion de habilitar la
modificacion. Si la edicion se completa sin problemas de validacion, se
pregunta si desea guardar los cambios. Si se acepta, se guarda en la
base de datos la informacion modificada y se actualiza el JWT de la sesion.
- La imagen aun no se modifica.
- Agregados metodos en la pagina Encuesta para evitar excepciones al subir
archivos.
- Modificadas las clases ws y las paginas de php necesarias para realizar
los nuevos pedidos.

+++ Pizzeria Argenta SRL/src/app/components/menu/menu.component.ts
@@ -16,9 +16,12 @@ export class MenuComponent implements OnInit {
 
   salir()
   {
-    localStorage.setItem('token', null);
-    window.alert('Chauuuuuu!!!');
-    this.router.navigate(['/login']);
+    if (confirm("¿Desea cerrar la sesion actual?"))
+    {
+      localStorage.setItem('token', null);
+      window.alert('Sesion cerrada!!!');
+      this.router.navigate(['/login']);
+    }
   }
 
   Comprobar()
